Restructure and change Standard Schema spec
